Fix API 19 drawable crash in ManageGroupFragment.
