--- app/src/main/res/drawable/group_resource_placeholder_80.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="oval">
-            <solid android:color="?attr/resource_placeholder_background" />
-        </shape>
-    </item>
-    <item
-        android:bottom="24dp"
-        android:drawable="@drawable/ic_group_80"
-        android:left="24dp"
-        android:right="24dp"
-        android:top="24dp" />
-
-    <item android:drawable="?attr/resource_placeholder_gradient" />
-</layer-list>
