+++ app/src/main/res/values/attrs.xml
@@ -31,9 +31,7 @@
     <attr name="conversation_list_compose_icon_tint" format="color" />
     <attr name="conversation_list_camera_button_background" format="color"/>
 
-    <attr name="group_resource_placeholder_80" format="reference" />
     <attr name="resource_placeholder_gradient" format="reference" />
-    <attr name="resource_placeholder_background" format="color" />
 
     <attr name="avatar_selection_take_photo" format="reference" />
     <attr name="avatar_selection_pick_photo" format="reference" />
